debugger: rename inspector-cli test module to debugger
PR-URL: https://github.com/nodejs/node/pull/38530
