+++ test/common/README.md
@@ -9,6 +9,7 @@ This directory contains modules used to test the Node.js implementation.
 * [Common module API](#common-module-api)
 * [Countdown module](#countdown-module)
 * [CPU Profiler module](#cpu-profiler-module)
+* [Debugger module](#debugger-module)
 * [DNS module](#dns-module)
 * [Duplex pair helper](#duplex-pair-helper)
 * [Environment variables](#environment-variables)
@@ -16,7 +17,6 @@ This directory contains modules used to test the Node.js implementation.
 * [Heap dump checker module](#heap-dump-checker-module)
 * [hijackstdio module](#hijackstdio-module)
 * [HTTP2 module](#http2-module)
-* [Inspector CLI module](#inspector-cli-module)
 * [Internet module](#internet-module)
 * [ongc module](#ongc-module)
 * [Report module](#report-module)
@@ -506,6 +506,34 @@ Sampling interval in microseconds.
 Throws an `AssertionError` if there are no call frames with the expected
 `suffix` in the profiling data contained in `file`.
 
+## Debugger module
+
+Provides common functionality for tests for `node inspect`.
+
+### `startCLI(args[[, flags], spawnOpts])`
+
+* `args` [&lt;string>][]
+* `flags` [&lt;string>][] default = []
+* `showOpts` [&lt;Object>][] default = {}
+* return [&lt;Object>][]
+
+Returns a null-prototype object with properties that are functions and getters
+used to interact with the `node inspect` CLI. These functions are:
+
+* `flushOutput()`
+* `waitFor()`
+* `waitForPrompt()`
+* `waitForInitialBreak()`
+* `breakInfo`
+* `ctrlC()`
+* `output`
+* `rawOutput`
+* `parseSourceLines()`
+* `writeLine()`
+* `command()`
+* `stepCommand()`
+* `quit()`
+
 ## `DNS` Module
 
 The `DNS` module provides utilities related to the `dns` built-in module.
@@ -823,34 +851,6 @@ upon initial establishment of a connection.
 ```js
 socket.write(http2.kClientMagic);
 ```
-
-## Inspector CLI module
-
-Provides common functionality for tests for `node inspect`.
-
-### `startCLI(args[[, flags], spawnOpts])`
-
-* `args` [&lt;string>][]
-* `flags` [&lt;string>][] default = []
-* `showOpts` [&lt;Object>][] default = {}
-* return [&lt;Object>][]
-
-Returns a null-prototype object with properties that are functions and getters
-used to interact with the `node inspect` CLI. These functions are:
-
-* `flushOutput()`
-* `waitFor()`
-* `waitForPrompt()`
-* `waitForInitialBreak()`
-* `breakInfo`
-* `ctrlC()`
-* `output`
-* `rawOutput`
-* `parseSourceLines()`
-* `writeLine()`
-* `command()`
-* `stepCommand()`
-* `quit()`
 
 ## Internet Module
