+++ test/inspector-cli/test-inspector-cli-auto-resume.js
@@ -4,7 +4,7 @@ const common = require('../common');
 common.skipIfInspectorDisabled();
 
 const fixtures = require('../common/fixtures');
-const startCLI = require('../common/inspector-cli');
+const startCLI = require('../common/debugger');
 const { addLibraryPath } = require('../common/shared-lib-util');
 
 const assert = require('assert');
