+++ test/inspector-cli/test-inspector-cli-low-level.js
@@ -4,7 +4,7 @@ const common = require('../common');
 common.skipIfInspectorDisabled();
 
 const fixtures = require('../common/fixtures');
-const startCLI = require('../common/inspector-cli');
+const startCLI = require('../common/debugger');
 const assert = require('assert');
 
 // Debugger agent direct access.
